refactor: simplify Chat transcript row policies
