+++ apps/desktop/src/features/chat/transcriptRows/firstPresent.ts
@@ -1,0 +1,8 @@
+export function firstPresent(...values: readonly (string | null | undefined)[]): string {
+  for (const value of values) {
+    if (value !== undefined && value !== null && value.trim().length > 0) {
+      return value;
+    }
+  }
+  return "";
+}
