+++ apps/desktop/src/features/chat/transcriptRows/index.ts
@@ -1,4 +1,4 @@
-export { isAskQuestionToolRow } from "./transcriptRowProjector";
-export type { TranscriptAskQuestionToolRow } from "./transcriptRowProjector";
+export { isAskQuestionToolRow } from "./transcriptAskQuestionPolicy";
+export type { TranscriptAskQuestionToolRow } from "./transcriptAskQuestionPolicy";
 export { TranscriptNoticeReviewerSlot } from "./TranscriptNoticeReviewerSlot";
 export { TranscriptAskQuestionRow as TranscriptAskQuestionSlot } from "./TranscriptAskQuestionRow";
