+++ apps/desktop/src/features/chat/transcriptRows/transcriptReviewerPolicy.ts
@@ -1,0 +1,3 @@
+export function reviewerFeedbackCopyText(suggestions: readonly string[]): string {
+  return suggestions.map((suggestion, index) => `${String(index + 1)}. ${suggestion}`).join("\n");
+}
